requirements.txt
# Airflow
apache-airflow==2.8.1
apache-airflow-providers-postgres==5.10.0

# Data Processing
pandas==2.1.4
numpy==1.24.3
pyarrow==14.0.2
fastparquet==2024.2.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy

# Machine Learning
scikit-learn==1.3.2
scipy==1.11.4

# Visualization
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0

pyspark==3.5.0

# HTTP Requests
requests==2.31.0
urllib3==2.1.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
openpyxl

# Testing
pytest==7.4.3
pytest-cov==4.1.0